+++ src/components/ui/alert-dialog/index.ts
@@ -1,5 +1,4 @@
 import { AlertDialog as AlertDialogPrimitive } from "@/lib";
-
 const Root = AlertDialogPrimitive.Root;
 const Trigger = AlertDialogPrimitive.Trigger;
 
@@ -12,6 +11,7 @@ import Header from "./alert-dialog-header.svelte";
 import Overlay from "./alert-dialog-overlay.svelte";
 import Content from "./alert-dialog-content.svelte";
 import Description from "./alert-dialog-description.svelte";
+
 import type { TransitionConfig } from "svelte/transition";
 import { cubicOut } from "svelte/easing";