+++ src/lib/internal/index.ts
@@ -1,2 +1,3 @@
+export * from "./is.js";
 export * from "./helpers.js";
 export * from "./types.js";