+++ src/utils/index.ts
@@ -1,5 +1,4 @@
 export * from "./styles";
 export * from "./copy-code";
-export * from "./is";
 export * from "./markdown";
 export * from "./styles";